Tests/reorganize: move setWordWrap tests to own file
